+++ src/ids-popup-menu/ids-popup-menu.js
@@ -52,11 +52,8 @@ class IdsPopupMenu extends mix(IdsMenu).with(IdsRenderLoopMixin, IdsEventsMixin)
    * @returns {string} The template
    */
   template() {
-    return `<ids-popup class="ids-popup-menu" type="menu">
-      <nav class="ids-menu" slot="content">
-        <slot></slot>
-      </nav>
-    </ids-popup>`;
+    const menuTemplate = IdsMenu.prototype.template.apply(this);
+    return `<ids-popup class="ids-popup-menu" type="menu">${menuTemplate}</ids-popup>`;
   }
 
   /**
@@ -94,7 +91,7 @@ class IdsPopupMenu extends mix(IdsMenu).with(IdsRenderLoopMixin, IdsEventsMixin)
       this.rl.register(new IdsRenderLoopItem({
         duration: 1,
         timeoutCallback: () => {
-          this.focusTarget.focus();
+          this.focusTarget?.focus();
         }
       }));
     };